refactor(evaluations): split pure codegen from server-only validation
Extract the judge script template (wrapPromptAsEvaluationScript +
EVALUATION_CONVERSATION_PLACEHOLDER) and the sandbox-backed validators
into dedicated modules so compileSettingsToScript's import graph is
browser-safe, ahead of exposing it through the package browser entry.
