+++ packages/domain/evaluations/src/use-cases/alignment/generate-baseline-draft.ts
@@ -2,8 +2,8 @@ import { hashOptimizationCandidateText } from "@domain/optimizations"
 import { Effect } from "effect"
 import { generateBaselinePromptText } from "../../alignment/baseline-prompt.ts"
 import type { GeneratedEvaluationDraft } from "../../alignment/types.ts"
+import { wrapPromptAsEvaluationScript } from "../../codegen/judge-script-template.ts"
 import { defaultEvaluationTrigger } from "../../entities/evaluation.ts"
-import { wrapPromptAsEvaluationScript } from "../../runtime/evaluation-execution.ts"
 
 // TODO(eval-sandbox): restore LLM-based baseline generation for arbitrary scripts when sandbox
 // is available.
