+++ packages/domain/evaluations/src/use-cases/preview-evaluation.ts
@@ -20,7 +20,8 @@ import {
 } from "@domain/spans"
 import { Effect } from "effect"
 import { z } from "zod"
-import { compileSettingsToScript, validateEvaluationScriptCompiles } from "../codegen/compile-settings-to-script.ts"
+import { compileSettingsToScript } from "../codegen/compile-settings-to-script.ts"
+import { validateEvaluationScriptCompiles } from "../codegen/validate-evaluation-script.ts"
 import { loadScriptSessionContext } from "../runtime/load-session-context.ts"
 import { executeEvaluationScriptSandboxed } from "../runtime/sandbox-execution.ts"
 import { buildSemanticSimilarityHost } from "../runtime/semantic-similarity.ts"
